+++ src/name_srvc_pckt.c
@@ -760,7 +760,7 @@ inline enum name_srvc_rdata_type name_srvc_understand_resource(uint16_t rrtype,
 
 
 void *master_name_srvc_pckt_reader(void *packet,
-				   int len,
+				   long len,
 				   uint16_t *tid) {
   struct name_srvc_packet *result;
   struct name_srvc_question_lst *cur_qstn;
@@ -924,7 +924,7 @@ void *master_name_srvc_pckt_reader(void *packet,
 }
 
 void *master_name_srvc_pckt_writer(void *packet_ptr,
-				   unsigned int *pckt_len,
+				   unsigned long *pckt_len,
 				   void *packet_field,
 				   unsigned char transport) {
   struct name_srvc_packet *packet;
